% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_functions.R
\name{annotate_cell_types}
\alias{annotate_cell_types}
\title{Rename clusters with cell type labels}
\usage{
annotate_cell_types(seurat_obj, labels)
}
\arguments{
\item{seurat_obj}{A Seurat object with clusters.}

\item{labels}{A character vector of new cluster names.}
}
\value{
A Seurat object with renamed identities.
}
\description{
Renames cluster IDs with user-provided cell type annotations.
}
