% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_functions.R
\name{filter_protein_coding_genes}
\alias{filter_protein_coding_genes}
\title{Filter protein-coding genes from a GTF file and subset a Seurat object}
\usage{
filter_protein_coding_genes(seurat_obj, gtf_path)
}
\arguments{
\item{seurat_obj}{A Seurat object}

\item{gtf_path}{Path to the GTF file}
}
\value{
A filtered Seurat object containing only protein-coding genes
}
\description{
Filter protein-coding genes from a GTF file and subset a Seurat object
}
