% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03.function_filter_unwanted_genes.R
\name{filter_unwanted_genes}
\alias{filter_unwanted_genes}
\title{Filter ribosomal, mitochondrial and pseudogenes from a Seurat object}
\usage{
filter_unwanted_genes(
  seurat_obj,
  out_list_file = "removed_genes_list.txt",
  out_summary_file = "removed_genes_summary.txt"
)
}
\arguments{
\item{seurat_obj}{A Seurat object}

\item{out_list_file}{Path to save the list of removed gene names (default: "removed_genes_list.txt")}

\item{out_summary_file}{Path to save a summary table of removed genes by category (default: "removed_genes_summary.txt")}
}
\value{
A filtered Seurat object
}
\description{
This function removes ribosomal genes, ribosomal pseudogenes, and mitochondrial genes from a Seurat object
based on gene name patterns (e.g. "RPS", "RPL", "MT-").
}
