% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_functions.R
\name{find_top_markers}
\alias{find_top_markers}
\title{Find top marker genes per cluster}
\usage{
find_top_markers(seurat_obj)
}
\arguments{
\item{seurat_obj}{A clustered Seurat object.}
}
\value{
A dataframe of top markers.
}
\description{
Identifies the most highly expressed gene per cluster.
}
