% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_functions.R
\name{infer_tissue_origin}
\alias{infer_tissue_origin}
\title{Infer tissue origin from cell types}
\usage{
infer_tissue_origin(seurat_obj)
}
\arguments{
\item{seurat_obj}{A Seurat object with annotated identities.}
}
\value{
A string hypothesis for the tissue of origin.
}
\description{
Based on annotated cell types, returns a guess for tissue origin.
}
