% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01.function_select_protein_coding_genes.R
\name{select_protein_coding_genes}
\alias{select_protein_coding_genes}
\title{Select protein-coding genes from a Seurat object using a GTF annotation}
\usage{
select_protein_coding_genes(seurat_obj, gtf_path)
}
\arguments{
\item{seurat_obj}{A Seurat object created from 10X data}

\item{gtf_path}{Path to the GTF annotation file (e.g., from Ensembl)}
}
\value{
A filtered Seurat object containing only protein-coding genes
}
\description{
This function filters a Seurat object to retain only protein-coding genes based on a GTF annotation file.
}
