% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_functions.R
\name{filter_unwanted_genes}
\alias{filter_unwanted_genes}
\title{Filter mitochondrial, ribosomal and ribosomal pseudogenes}
\usage{
filter_unwanted_genes(seurat_obj)
}
\arguments{
\item{seurat_obj}{A Seurat object.}
}
\value{
A list with two elements: (1) filtered Seurat object, and (2) summary table of removed gene counts.
}
\description{
Removes genes belonging to mitochondrial, ribosomal and pseudogene categories.
}
